+++ hpobench/container/benchmarks/nas/nasbench_101.py
@@ -10,6 +10,7 @@ class NASCifar10ABenchmark(AbstractBenchmarkClient):
     def __init__(self, **kwargs):
         kwargs['benchmark_name'] = kwargs.get('benchmark_name', 'NASCifar10ABenchmark')
         kwargs['container_name'] = kwargs.get('container_name', 'nasbench_101')
+        kwargs['latest'] = kwargs.get('container_tag', '0.0.3')
         super(NASCifar10ABenchmark, self).__init__(**kwargs)
 
 
@@ -17,6 +18,7 @@ class NASCifar10BBenchmark(AbstractBenchmarkClient):
     def __init__(self, **kwargs):
         kwargs['benchmark_name'] = kwargs.get('benchmark_name', 'NASCifar10BBenchmark')
         kwargs['container_name'] = kwargs.get('container_name', 'nasbench_101')
+        kwargs['latest'] = kwargs.get('container_tag', '0.0.3')
         super(NASCifar10BBenchmark, self).__init__(**kwargs)
 
 
@@ -24,4 +26,5 @@ class NASCifar10CBenchmark(AbstractBenchmarkClient):
     def __init__(self, **kwargs):
         kwargs['benchmark_name'] = kwargs.get('benchmark_name', 'NASCifar10CBenchmark')
         kwargs['container_name'] = kwargs.get('container_name', 'nasbench_101')
+        kwargs['latest'] = kwargs.get('container_tag', '0.0.3')
         super(NASCifar10CBenchmark, self).__init__(**kwargs)